requirements.txt
fastapi
uvicorn
kafka-python
mysql-connector-python
redis